refactor cmd/nerdctl for Linux
